Fix burn-in test UI overlapped by bootup animation
Symptom:
1. Burn-in test UI is overlapped by bootup animation and the disclaimer
   page.
2. The burn-in test time does not blink.

Root Cause:
- The bootup animation is displayed without checking if the burn-in test
  is ongoing.

Fix:
1. Add Enum::OperationMode.PRODUCTION_TEST, and set to this mode when the
   burn-in test started.
2. Not display the bootup animation when the current mode is
   Enum::OperationMode.PRODUCTION_TEST.

Testing and Verification Instructions:
1. Enable UNIT_TEST_BURNIN in TEST_pub.h.
2. Compile and run.
=> The burn-in test UI should be displayed normally.

Issue: Autooem-IXWW22-5745
Type: Bug Fix
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/84969
Tested-by: Huang, Kenny <kenny.huang@garmin.com>
Approve-Change: Liao, Edward <edward.liao@garmin.com>

+++ EmbeddedWizard/EW_pub.h
@@ -41,7 +41,7 @@ void EW_reset_to_factory_default( void );
 
 uint32_t EW_get_ccuid( void );
 uint16_t EW_get_qrcode_dummy( void );
-uint16_t EW_get_qrcode_passkey( void );
+uint32_t EW_get_qrcode_passkey( void );
 
 void EW_notify_navi_map_update( void );
 void EW_notify_engine_speed_changed( int speed );
